feat: versión 1.3.0 - nuevo sistema de gestión y mejoras en acciones especiales
- Implementado nuevo tipo de gestión 'time' para acciones especiales
- Agregado dato 'hits' en cálculos de daño (interno)
- Añadido daño máximo del giro en Impacto de Obús (Katsuki Bakugô)
- Refactorizada visualización de tablas para mejor legibilidad
- Migrado acciones 100% de gestión 'percent' a 'time'
- Actualizado daño en Golpe Metralla: individual → por tiro (Katsuki Bakugô)
- Ajustado porcentaje mínimo en Home Run Cometa (Ochako Uraraka)
- Corregidos datos de consumo en Detroit Smash Quíntuple (Izuku Midoriya OFA)

+++ src/personajes.h
@@ -41,7 +41,7 @@ typedef struct {
     char name[55];
     char *description;
     int typeGestion;
-    int shotsOrPercentUsage[9];
+    float shotsOrPercentUsage[9];
     float reload[9];
     float fullReload[9];
     int countDamages;
